requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-evals
python-dotenv
